operator-api/requirements.txt
fastapi
uvicorn[standard]
pyyaml
pydantic
snowflake-connector-python
requests
datahub-emitter